+++ tests/backends/base/playback.py
@@ -434,18 +434,18 @@ class PlaybackControllerTest(object):
 
     @populate_tracklist
     def test_previous_track_before_play(self):
-        self.assertEqual(self.playback.track_at_previous, None)
+        self.assertEqual(self.playback.tl_track_at_previous, None)
 
     @populate_tracklist
     def test_previous_track_after_play(self):
         self.playback.play()
-        self.assertEqual(self.playback.track_at_previous, None)
+        self.assertEqual(self.playback.tl_track_at_previous, None)
 
     @populate_tracklist
     def test_previous_track_after_next(self):
         self.playback.play()
         self.playback.next()
-        self.assertEqual(self.playback.track_at_previous, self.tracks[0])
+        self.assertEqual(self.playback.tl_track_at_previous, self.tl_tracks[0])
 
     @populate_tracklist
     def test_previous_track_after_previous(self):
@@ -453,10 +453,10 @@ class PlaybackControllerTest(object):
         self.playback.next()  # At track 1
         self.playback.next()  # At track 2
         self.playback.previous()  # At track 1
-        self.assertEqual(self.playback.track_at_previous, self.tracks[0])
+        self.assertEqual(self.playback.tl_track_at_previous, self.tl_tracks[0])
 
     def test_previous_track_empty_playlist(self):
-        self.assertEqual(self.playback.track_at_previous, None)
+        self.assertEqual(self.playback.tl_track_at_previous, None)
 
     @populate_tracklist
     def test_previous_track_with_consume(self):
@@ -464,7 +464,8 @@ class PlaybackControllerTest(object):
         for _ in self.tracks:
             self.playback.next()
             self.assertEqual(
-                self.playback.track_at_previous, self.playback.current_track)
+                self.playback.tl_track_at_previous,
+                self.playback.current_tl_track)
 
     @populate_tracklist
     def test_previous_track_with_random(self):
@@ -472,7 +473,8 @@ class PlaybackControllerTest(object):
         for _ in self.tracks:
             self.playback.next()
             self.assertEqual(
-                self.playback.track_at_previous, self.playback.current_track)
+                self.playback.tl_track_at_previous,
+                self.playback.current_tl_track)
 
     @populate_tracklist
     def test_initial_current_track(self):